refactor(capabilities): host the tcp cap actor on its struct
Move the tcp capability's NativeActor impl into runtime.rs under #[runtime]
and lift #[actor(singleton)] onto the TcpCapability struct (ADR-0123). Issue #2366.
